"lakefs setup" to set up ACL groups
Depending on auth.ui_config.rbac "simplified" or "external", create
appropriate initial groups.
